[IOTMID-121] Suport for generic devices
This commit implements the support for generic devices.
In order to use them, the device manufacturer should
add a new file in './translators' folder containing
a translator function implementation. It should convert
the data from the device to something like this:

{
    "param-1" : "value-1",
    "param-2" : "value-2"
}

Also, when adding a new service, the user should
indicate which device is being added (by setting
the 'type' request data), such as:
{
    "services": [
    {
        "resource": "/myHome",
        "apikey": "transmitter",
        "type": "SensorWebDevice",
    }
    ]
}

Note that the 'type' parameter is the same name as the translator to be used.

When creating the device, the user will need to set
internal_attributes so that it will contain the topics
for configuration and attributes. An example:
"devices": [
{
    "device_id": "sensorweb-04",
    "entity_name": "sweb-4",
    "entity_type": "SensorWebDevice",
    "attributes": [
	  { "object_id": "t", "name": "Temperature", "type": "celsius" },
	  { "object_id": "l", "name": "Luminosity", "type": "lumens" }
    ],
    "transport" : "mqtt",
    "internal_attributes":  {
	"attributes" : [
	    { "topic" : "tcp:mqtt:/sweb-4/publish" }
	],
	"configuration" : [
	    { "topic" : "tcp:mqtt:/sweb-4/configuration/+" }
	]
    }
}
]

Change-Id: I0eb0f6b0931bc19c7e775dbf9b17c5c4125eb06d