[IOTMID-224] Support for device accessibility status
This commit implements the support for device accessibility
status. The status is based on how many messages per second
a particular device generates and a few extra parameters, such
as how many message could be skipped before deeming that device
as offline.

Also, the user might specify which parameters should be used to
detect that a devide went offline. The internal_attributes
parameter can be used when creating the device:

"internal_attributes":  {
    "timeout" : {
        "sampleQueueMaxSize" : 4,
        "waitMultiplier" : 3 ,
        "minimumTimeoutBase" : 50
    }
}

Or the user can explicitly specify the period of time between
messages:
"internal_attributes":  {
    "timeout" : {
        "periodicity" : 2000,
        "waitMultiplier" : 3
    }
}

The 'periodicity' attribute is in miliseconds. In the future
we will accept other timescales.

Change-Id: Ibd8c779604e1d859d8e3dc6085cb458f0f294a6b