This commit reimplements iotagent-json
This new code is intended to be simpler (and hopefully faster)
than the original one. It doesn't require any extra configuration
(such as creating devices) - for each new message that it receives,
it asks (or, at least, will ask) DeviceManager if that device exists
and which is its ID (and related fiware-service, for now). After
receiving a valid answer (the device indeed exists), it sends
an update message to Orion, just like before (almost - it now uses
the simpler Orion's v2 API)